Move constants into relevant files and refactored main game input loop
Constants was moved to their relevant files to make it easier to
organize an increasing number of constants. It also makes it more
readable when refering to it: we already know that the variable is
a constant in `constants::`. Putting it into the files more relevant
makes it easier to tell what the constant is about.

Changed `loop {match prompt()}` to `while let` to reduce the amount of
indentation in the main game loop.