Massive refactoring of tile's data and game actions in the main loop
Each tile struct no longer stores the JSON serde_json::Value object.
Instead it only stores the information that is needed as defined in
the struct's field. The tiles parse the JSON data on creation.

The main game loop has been abstracted. Monopoly object stores all
the objects needed for the game except for players because some
code related to players is needed in `main.rs` in addition to the
Monopoly object. To keep players non-global, players are kept in
`main.rs`. Monopoly was created to keep the main loop simplified
and to do all the game actions/rules/logic while the `main.rs` only
runs the functions as necessary based on input and display output
messages.

Also, purchasing of tiles to complete or improve set ownership
now increases rent for all other tiles of the same set.