Added JSON validator, so we can assume certain keys exist
This reduces the amount of `.get(key).expect(error::ERROR_MESSAGE)`,
and the whole nest of `.get()`s if trying to fetch nested objects
which makes it more readable. It doesn't even stop the error. It still
panics, but with a better error message of where the error occured.
These error messages are no longer needed if they are checked once
at the very start of the program with clearer messages of where and why
the error occured.