Moved players to Monopoly struct so it can fully handle all game actions
It was poor design for players to be created in `main.rs` while
most of the game actions were inside `game.rs::Monopoly`. It had to
constantly refer to players by ID, thus not fully being able to
do all game actions.

Now, that `main.rs` cannot interact with players, it can only refer
to it by ID or by requesting it through a getter in `Monopoly`.